[refactor] 채팅방 참여자 수를 participant 집계 기반으로 전환

+++ src/main/java/_ganzi/codoc/chat/dto/UserChatRoomListQueryResult.java
@@ -6,6 +6,6 @@ public record UserChatRoomListQueryResult(
         Long participantId,
         Long roomId,
         String title,
-        int participantsCount,
+        long participantsCount,
         String lastMessagePreview,
         Instant lastMessageAt) {}
